+++ NEWS.md
@@ -1,6 +1,9 @@
 # Version (development version)
 
- * ...
+## Deprecated and Defunct
+
+ * The 'multiprocess' strategy has now been fully removed.  Please use
+   'multisession' (recommended) or 'multicore' instead.
 
 
 # Version 1.32.0 [2023-03-06]